Leave index unchanged and add link to my page
