+++ src/pages/index.tsx
@@ -1,20 +1,10 @@
 import * as React from 'react';
-import { Board } from '../components/board';
-import { css } from '../../node_modules/emotion';
-
-const Game = () => (
-  <div className={css({
-    display: 'flex',
-    flexDirection: 'row',
-    textAlign: 'center',
-  })}>
-    <div>
-      <p>Try some sample rules from <a href="https://en.wikipedia.org/wiki/Life-like_cellular_automaton">here</a></p>
-      <Board />
-    </div>
-  </div>
-);
+import Link from 'gatsby-link';
 
 export default () => (
-  <Game />
+  <div>
+    <h1>Hi people</h1>
+    <Link to="/page-2/">Go to page 2</Link>
+    <Link to="/psrh/">Go to Game of Life</Link>
+  </div>
 );
